extract logits filtering into dedicated SRP microcrate
